qualifier changes
