+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/auto/roadrunnerauto/AutoRedGoal3Simple.java
@@ -3,6 +3,7 @@ package org.firstinspires.ftc.teamcode.opmodes.auto.roadrunnerauto;
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.pedropathing.follower.Follower;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.teamcode.modules.Intake;
@@ -10,6 +11,7 @@ import org.firstinspires.ftc.teamcode.modules.Shooter;
 import org.firstinspires.ftc.teamcode.modules.drivetrainrr.DriveTrainMecanum;
 import org.firstinspires.ftc.teamcode.pedroPathing.Constants;
 
+@Disabled
 @Autonomous(name = "RED Goal 3 Artifacts", group = "1")
 public class AutoRedGoal3Simple extends LinearOpMode {
