+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/auto/roadrunnerauto/AutoRedHumanPark.java
@@ -2,11 +2,13 @@ package org.firstinspires.ftc.teamcode.opmodes.auto.roadrunnerauto;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.teamcode.modules.drivetrainrr.DriveTrainMecanum;
 
 @Autonomous(name = "RED Human Park", group = "RedPark")
+@Disabled
 public class AutoRedHumanPark extends LinearOpMode {
 
     @Override
